Rename utils move some deploy functionality in there
